feat: move NostrProfile to feature
